Completely remove OU Noise.
One of the features of the MixSig project is the ability to generate unevenly sampled time series data.

Since OU Noise is calculated based on fixed interval timesteps, it doesn't make much sense to use it here.

If we want noise, there are other, more appropriate ways to go about it.