Add sequence_code attribute to MixedSignal class.
With this, sequence_type is now a property that is defined by the sequence_code.

Update the sliding and boxcar generators to use the sequence_code.

Allow sequence_code to be passed into generators for to facilitate unit testing.

Because of this, the generate() function is now in charge of setting the X and y variables.

Remove the useless _generate() function and merge its contents into generator()

Add full coverage unit tests for sliding and boxcar generators.