Fix: Put all Wave instances in a list so it is easier to iterate over them.
Before, the waves belonging to MixedWave were not easily accessable.

This also meant Refactoring and some manual rollback.

This also changes the API. Don't nest wave objects inside the mixed_wave object.
Instead, pass them all in through a list.  The code will work out which is which.
Now, you need specify which entry is the mixed_wave by setting 'name=mixed_wave'.
Also, If a wave isn't the mixed_wave and doesn't have a 'time' parameter,
Then we will assume that it will be absorbed into the main mixed_wave and not,
be run standalone.

 - Add a classification_type to help us toggle the shape of the outputs.
 - Remove the is_independent flag from waves.
   There is an easier way to keep track of this.
 -